requirements.txt
# Ultra-Efficient XRP Trading Bot Dependencies
# Core dependencies for the trading system

# Hyperliquid SDK
hyperliquid-python-sdk

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Machine Learning
torch>=1.9.0
stable-baselines3>=1.5.0
gymnasium>=0.26.0

# Performance optimization
numba>=0.56.0
psutil>=5.8.0

# Async and networking
aiohttp>=3.8.0
websockets>=10.0

# Configuration and utilities
python-dotenv>=0.19.0
pyyaml>=6.0

# Logging and monitoring
colorlog>=6.6.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.18.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Telegram notifications
python-telegram-bot>=13.0

# Optional: Database support
sqlite3  # Built-in with Python

# Optional: Advanced ML
optuna>=3.0.0  # Hyperparameter optimization
cloudpickle>=2.0.0  # Serialization
llvmlite>=0.38.0  # Numba dependency
